% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.lme1en.R
\name{cv.lme1en}
\alias{cv.lme1en}
\title{cv.lme1en}
\usage{
cv.lme1en(y, X, batch, rho = 0, lambda = 1, alpha = 0.5,
  beta = NULL, glmnetPenalty = TRUE, glmnetWarmup = TRUE,
  maxit = 10000, toler = 1e-05, verbose = FALSE, errorFun = NULL,
  Kfold = 10, CViter = 10, parallel = TRUE, seed = NULL)
}
\arguments{
\item{y}{response vector. Must be normalized before input.}

\item{X}{Design matrix belonging to fixed effects coefficients (beta). Must be normalized before input.}

\item{batch}{factor with batch effect names (vector for each observations)}

\item{rho}{proportion of variation explained by batch effect}

\item{lambda}{numeric, penalty levels for fixed effects betas (component 1)}

\item{alpha}{numeric, penalty levels for fixed effects betas (balancing LASSO/RIDGE erros)}

\item{beta}{numeric, initial values of the beta coefficients (using glmnet or marginal estimates if not provided)}

\item{glmnetPenalty}{boolean, whether to use the original peanalty (FALSE) or the glmnet penalty (TRUE)}

\item{glmnetWarmup}{boolean, whether to use glmnet beta-estimates as warmup (if not marginals are used)}

\item{maxit}{maximum number of iterations (i.e. forloops) in the coordinate decent algorithm}

\item{toler}{tolerance level of beta changes for each iterations (similar to 'thresh' in glmnet)}

\item{verbose}{boolean, show progress Default: FALSE}

\item{errorFun}{function, The user can provide own error function to perform cross validation with (Default is MSE if function not provided). The function must be able to take equally long two vectors and return a number}

\item{Kfold}{The train-test fold split: The training proportion becomes '(Kfold-1)/Kfold'}

\item{CViter}{Number of iterations of the cross-validation}

\item{parallel}{Boolean of whether parallelization in CV process should be applied (utilizing all CPUs)}

\item{seed}{The user can set seed for reproducibility if wanted}
}
\value{
matrix with error values (test and train) for each iterations
}
\description{
Performing crossvalidation of lme1en (extension of inputs)
}
\details{
The function performs (possibly parallelized) K-fold cross validation for given alpha,lambda,rho
The function assume that the response y and design matrix X are centralized (no intercept returned from function)
}
\examples{
\dontrun{ 
dat = genData(seed=1)
CVerrors = cv.lme1en(y=dat$y,X=dat$X,batch=dat$batch, rho=0.3,lambda=0.1,alpha=0.5,verbose=TRUE,seed=1)
}
}
