% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genData}
\alias{genData}
\title{genData}
\usage{
genData(ntot = 1000, nsites = 1000, nbatches = 5, propAgeAs = 0.05,
  sd_batch = 0.5, sd_signal = 0.1, sd_bparam = 0.03,
  ageRange = c(10, 35), seed = NULL, verbose = TRUE)
}
\arguments{
\item{ntot}{total number of observations}

\item{nsites}{number of sites}

\item{nbatches}{number of batches}

\item{propAgeAs}{#proportion of sites with assosiation to age}

\item{sd_batch}{standard dev of batch effects (normal distr effects)}

\item{sd_signal}{standard dev of signal (norm distr noise)}

\item{sd_bparam}{sd of age coefficient}

\item{ageRange}{Age range of considered ages}

\item{seed}{The user can set seed if wanted}

\item{verbose}{boolean, show progress Default: FALSE}
}
\value{
A list with variables y (response), X (design matrix belonging to fixed effects coefficients), batch (factor with batch effect names), age (true ages)
}
\description{
Generating a example dataset
}
\details{
The function generates a dataset based on beta distribution where age (uniformly distributed) is a underlying explanatory variable
}
\examples{
\dontrun{ 
dat = genData(seed=1)
}
}
