% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme1en.R
\name{lme1en}
\alias{lme1en}
\title{lme1en}
\usage{
lme1en(y, X, batch, lambda = 1, alpha = 0.5, rho = 0, beta = NULL,
  glmnetPenalty = TRUE, glmnetWarmup = TRUE, maxit = 10000,
  toler = 1e-05, verbose = FALSE)
}
\arguments{
\item{y}{response vector. Should be standardized before input (use scale).}

\item{X}{Design matrix belonging to fixed effects coefficients (beta). Should be standardized before input (use scale).}

\item{batch}{factor with batch effect names (vector for each observations)}

\item{lambda}{numeric, penalty levels for fixed effects betas (component 1)}

\item{alpha}{numeric, penalty levels for fixed effects betas (balancing LASSO/RIDGE erros)}

\item{rho}{#proportion of variation explained by batch effect}

\item{beta}{numeric, initial values of the beta coefficients (using glmnet or marginal estimates if not provided)}

\item{glmnetPenalty}{boolean, whether to use the original peanalty (FALSE) or the glmnet penalty (TRUE)}

\item{glmnetWarmup}{boolean, whether to use glmnet beta-estimates as warmup (if not marginals are used)}

\item{maxit}{maximum number of iterations (i.e. forloops) in the coordinate decent algorithm}

\item{toler}{tolerance level of beta changes for each iterations (similar to 'thresh' in glmnet)}

\item{verbose}{boolean, show progress Default: FALSE}
}
\value{
fitted beta values
}
\description{
Fitting Elastic Net with random intercept
}
\details{
The model is fitted using the coordinate decent algorithm (exact solutions)
The extended model includes the rho parameter (proportion of total variation) as argument. 
Calls a algorithm (iterate) in C which uses exact solutions of marginal beta's based on formula x = sgn(c/a)(|c/a| - b/a)_+, where ax + b*sgn(x) = c
The function assume that the response y and design matrix X are centralized (no intercept returned from function)
}
\examples{
\dontrun{ 
dat = genData(seed=1)
bhat = lme1en(y=dat$y,X=dat$X,batch=dat$batch, rho=0.3,lambda=0.1,alpha=0.5,verbose=TRUE)
}
}
