% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdata.R
\name{plotdata}
\alias{plotdata}
\title{plotdata}
\usage{
plotdata(x, y, txt = "", yr = NULL, grp = NULL)
}
\arguments{
\item{x}{True values}

\item{y}{Predicted values}

\item{txt}{Main label to show}

\item{yr}{Range of values}

\item{grp}{A factor vector for grouped points}
}
\description{
Plotting predicted and true response values together
}
