% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runexample.R
\name{runexample}
\alias{runexample}
\title{runexample}
\usage{
runexample(TestTrainProp = 0.6, ntot = 500, nsites = 1000,
  nbatches = 20, propAgeAs = 0.05, sd_batch = 1, sd_signal = 1,
  sd_bparam = 0.1, ageRange = c(10, 35), seed = NULL,
  verbose = TRUE)
}
\arguments{
\item{TestTrainProp}{Proportion of data to be used for training (remaining used as test)}

\item{ntot}{total number of observations (including train and test data)}

\item{nsites}{number of sites}

\item{nbatches}{number of batches}

\item{propAgeAs}{#proportion of sites with assosiation to age}

\item{sd_batch}{standard dev of batch effects (normal distr effects)}

\item{sd_signal}{standard dev of signal (norm distr noise)}

\item{sd_bparam}{sd of age coefficient}

\item{ageRange}{Age range of considered ages}

\item{seed}{The user can set seed for reproducibility if wanted}

\item{verbose}{boolean, show progress Default: TRUE}
}
\description{
Fit elastic net models (without/with batch effects) for a generated example dataset
}
\details{
The function do following steps: 1) simulate dataset 2) Split data into train/test 3) For two comparing models we run cross-valid for calibration and obtain performance using test data
}
\examples{
\dontrun{ 
 runexample(ntot=500,nsites=1000,nbatches=20,verbose=TRUE,seed=1, sd_batch = 1, sd_signal = 1, sd_bparam = 0.1)
}
}
