% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showResult.R
\name{showResult}
\alias{showResult}
\title{showResult}
\usage{
showResult(true, pred, train, test, errorFun)
}
\arguments{
\item{true}{True values (x)}

\item{pred}{Predicted values (y)}

\item{train}{index of train data}

\item{test}{index of test data}

\item{errorFun}{A error function (must return av constant when sending two equally long vectors)}
}
\description{
For plotting true and predicted values for train/test data
}
\details{
Allow for scaled variables (will scale back to original scale)
}
